Load a test image to demonstrate image loading capability
This adds SFML 2.6.1 as a dependency and uses it to load a test
image file. The only SFML modules that are built are the Graphics
module (for image loading capability) and the Windows module
(must be included for the Graphics module). This also changes the
CMake to copy across a `data` directory from the source directory
to the binary directory where the executable is located, so that
the executable can have easy access to the normal location of the
test image file.