Add ability to convert image to grayscale using two methods
This adds the ability to convert an image to grayscale using the
average method or the weighted method.

The average method simply averages the three RGB channels of each
pixel while preserving the alpha channel. The weighted method tries
to be a bit smarter because not every channel contributes equally
to the final image as the 'green' channel is far more visible to
humans than the 'blue' channel for instance. The weighted method
applies weights to each channel to adjust for this, and these weights
are based on the same weights that OpenCV uses for its grayscale
conversion.